refactor: useScale を Panzoom モジュールに取り込んだ
関係性が強いので。
